+++ core-bundle/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### DEV
 
+ * Re-add the "reset selection" buttons to the picker (see #856).
  * Trigger all the callbacks in the toggleVisibility() methods (see #756).
  * Only execute the command scheduler upon the "contao_backend" and "contao_frontend" routes (see #736).
  * Correctly set the important part in "edit multiple" mode (see #839).